Use an NSInteger so that the calculation gets rounded.

+++ Simplenote/SPTextView.m
@@ -40,7 +40,7 @@
 }
 
 - (CGFloat)getAdjustedInsetX: (CGFloat)viewWidth {
-    CGFloat adjustedInset = (viewWidth - kMaxEditorWidth) / 2;
+    NSInteger adjustedInset = (viewWidth - kMaxEditorWidth) / 2;
     
     return adjustedInset + kMinEditorPadding;
 }
